feat: add option to exit after launch
Using this option fixes #595 by proxy of spawning a new instance on the
current desktop.

I imagine cleaning up Browserosaurus can also help conserve memory usage
on those systems with little of it, at the cost of startup time.

Keep in mind if this option is set during development, Browserosaurus
*will* exit and you'll have to restart it on the command line to use
it as the default HTTP handler; but in production this would seamlessly
spin up the app when the user clicks a link.

This returns a promise from `openApp` which could perhaps be used for
more extensive error handling in the future.